WIP: Cleanup token (BC: Table "tokens" is dropped then recreated!)
* Clean comments
* Clean repository
* BC: token "value" is now stored in "uuid" column instead of id (see explanations below)

TokenFactory was generating a "unique id", but the value were always overriden because of doctrine schema (type guid).
Storing UUID as Primary Key is not recommended, neither storing them as char.

This commit use Ramsey\Uuuid (instead of using Uuid generated by the SQL Platform); token PK is now a classic "auto-generated" ID.
The UUID value is still used as the token identifier for api consumers.